# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  amount: number;
  type: TransactionType;
  description: string;
  date: string;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  createdAt: string;
}

export type TransactionInputType =
  (typeof TransactionInputType)[keyof typeof TransactionInputType];

export const TransactionInputType = {
  income: "income",
  expense: "expense",
} as const;

export interface TransactionInput {
  amount: number;
  type: TransactionInputType;
  description: string;
  date: string;
  /** @nullable */
  categoryId?: number | null;
}

export interface Category {
  id: number;
  name: string;
  color: string;
}

export interface CategoryInput {
  name: string;
  color: string;
}

export interface FinancialSummary {
  balance: number;
  totalIncome: number;
  totalExpense: number;
  transactionCount: number;
}

export interface CategoryExpense {
  /** @nullable */
  categoryId: number | null;
  categoryName: string;
  color: string;
  total: number;
  percentage: number;
}

export interface MonthlySummary {
  month: number;
  year: number;
  income: number;
  expense: number;
}

export type ListTransactionsParams = {
  type?: ListTransactionsType;
  /**
   * @nullable
   */
  categoryId?: number | null;
  limit?: number;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
} as const;