# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Finance App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: categories
    description: Category operations
  - name: summary
    description: Summary/analytics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense]
          required: false
        - in: query
          name: categoryId
          schema:
            type: ["integer", "null"]
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /summary:
    get:
      operationId: getSummary
      tags: [summary]
      summary: Get financial summary (balance, total income, total expense)
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"

  /summary/by-category:
    get:
      operationId: getExpenseByCategory
      tags: [summary]
      summary: Get expense totals grouped by category
      responses:
        "200":
          description: Expense breakdown by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryExpense"

  /summary/monthly:
    get:
      operationId: getMonthlySummary
      tags: [summary]
      summary: Get income and expense totals per month for the last 6 months
      responses:
        "200":
          description: Monthly summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlySummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Transaction:
      type: object
      required: [id, amount, type, description, date, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        date:
          type: string
          format: date
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    TransactionInput:
      type: object
      required: [amount, type, description, date]
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        date:
          type: string
          format: date
        categoryId:
          type: ["integer", "null"]

    Category:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string

    CategoryInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string

    FinancialSummary:
      type: object
      required: [balance, totalIncome, totalExpense, transactionCount]
      properties:
        balance:
          type: number
        totalIncome:
          type: number
        totalExpense:
          type: number
        transactionCount:
          type: integer

    CategoryExpense:
      type: object
      required: [categoryId, categoryName, color, total, percentage]
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: string
        color:
          type: string
        total:
          type: number
        percentage:
          type: number

    MonthlySummary:
      type: object
      required: [month, year, income, expense]
      properties:
        month:
          type: integer
        year:
          type: integer
        income:
          type: number
        expense:
          type: number