# lib/api-zod/src/generated/types/categoryExpense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance App API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryExpense {
  /** @nullable */
  categoryId: number | null;
  categoryName: string;
  color: string;
  total: number;
  percentage: number;
}