# lib/api-zod/src/generated/types/financialSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance App API
 * OpenAPI spec version: 0.1.0
 */

export interface FinancialSummary {
  balance: number;
  totalIncome: number;
  totalExpense: number;
  transactionCount: number;
}