# lib/api-zod/src/generated/types/monthlySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance App API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlySummary {
  month: number;
  year: number;
  income: number;
  expense: number;
}