# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance App API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInputType } from "./transactionInputType";

export interface TransactionInput {
  amount: number;
  type: TransactionInputType;
  description: string;
  date: Date;
  /** @nullable */
  categoryId?: number | null;
}